ti-img-rogue-driver: resolve issues with fdinfo
Bump the srcrev to include a fix for events where the driver is
requested to show advanced information about file descriptors that are
associated with the device but are not currently associated with a
rendering task.

This should address the kernel fault in pvr_show_fdinfo associated with
fdinfo/lsof.

Signed-off-by: Randolph Sapp <rs@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>