optee: Update to 4.7.0 tag
Update optee_client, optee_examples, optee_os and optee_test to the
newer 4.7.0 tag from the previous 4.6.0 one.

Further: meta-arm applies 0003-optee-enable-clang-support patch to
optee_os. However, in a recent upstream commit [0], the line to which
the patch applies was removed. Therefore the patch is no longer
required. As a result, tell optee_os.bbappend file to remove it.

[0] https://github.com/OP-TEE/optee_os/commit/3d5c2dd8b59ae888434b3f86842828131f115d90

Signed-off-by: Suhaas Joshi <s-joshi@ti.com>