meta-ti-bsp: optee: Add flags to enable RPMB and PKCS#11
RPMB is a secure storage mechanism used to store data in a separate
partition of compliant storage devices such as eMMC, NVME etc. It is
provided by TEE's, including OP-TEE.

Add CFG_RPMB_FS=y to enable this feature.

Also add CFG_RPMB_WRITE_KEY=y. This option tells optee_os to generate
and write the Authentication Key to the storage device controller during
the first access. The alternative is that the user would have to do it
themselves using tools such as mmc-tools.

NOTE: For production builds, CFG_RPMB_WRITE_KEY should be set to `n`.
But for purposes of SDK demo, it is enabled.

Lastly: add CFG_PKCS11_TA=y. This option enables PKCS#11 API support in
the form of a Trusted Application. Copy this TA to the relevant
location.

Signed-off-by: Suhaas Joshi <s-joshi@ti.com>