meta-ti-bsp: optee: Add flags to enable RPMB and PKCS#11
RPMB is a secure storage mechanism used to store data in a separate
partition of compliant storage devices such as eMMC, NVME etc. It is
provided by TEE's, including OP-TEE.

Add the following build options to optee_os:
    * CFG_REE_FS=n -> disables the default REE_FS, this is so that
      RPMB can be be demonstrated
    * CFG_RPMB_FS=y -> enables the RPMB feature
    * CFG_RPMB_WRITE_KEY=y -> generates and automatically writes the RPMB
      key if not already written (see NOTE below).
    * CFG_PKCS11_TA=y -> enables PKCS#11 API support in form of a
      Trusted Application. This commit also copies this TA to the
      relevant location.

In optee_client, do the following:
    * Add RPMB_EMU=1 option. This is enabled by-default, but even so,
      enable it explicitly. This option makes tee-supplicant emulate
      RPMB instead of using the actual hardware. The actual hardware
      should be used consciously since the key, once written, cannot be
      re-programmed. But in the emulated flow, each reboot wipes the key
      off, since the "emulated RPMB" is just a portion of primary
      memory.
    * Copy libckteec library files to relevant locations.

NOTE:

CFG_RPMB_WRITE_KEY=y sends the key in *plain text* to the normal Linux
world. This might be OK in development environments, but it is a huge
security risk in production! Therefore, this option must always be
disabled in production images.

RPMB_EMU=1 emulates RPMB, instead of using the real one. With
CFG_RPMB_WRITE_KEY=y, it also prevents the key from being written to the
real device. This option, also, must be disabled (RPMB_EMU=0) in
production environment.

Both of these options are enabled for the purposes of demo'ing RPMB
examples only, but are unsuited for production.

Signed-off-by: Suhaas Joshi <s-joshi@ti.com>